refactor: Extract swipe navigation logic to a separate hook
