--- commit_message.txt
@@ -1,1 +1,0 @@
-feat: Implement swipe to navigate in single page mode
